+++ http/client.go
@@ -63,12 +63,9 @@ func NewService(addr, token string) (*Service, error) {
 			Addr:  addr,
 			Token: token,
 		},
-		BucketService:    &BucketService{Client: httpClient},
-		DashboardService: &DashboardService{Client: httpClient},
-		OrganizationService: &OrganizationService{
-			Addr:  addr,
-			Token: token,
-		},
+		BucketService:       &BucketService{Client: httpClient},
+		DashboardService:    &DashboardService{Client: httpClient},
+		OrganizationService: &OrganizationService{Client: httpClient},
 		UserService: &UserService{
 			Addr:  addr,
 			Token: token,