+++ ui/src/alerting/utils/vis.test.ts
@@ -1,9 +1,17 @@
 import {getCheckVisTimeRange} from 'src/alerting/utils/vis'
 
+const duration = 'duration' as 'duration'
 const TESTS = [
-  ['5s', {lower: 'now() - 1500s'}],
-  ['1m', {lower: 'now() - 300m'}],
-  ['1m5s', {lower: 'now() - 300m1500s'}],
+  ['5s', {type: duration, lower: 'now() - 1500s', upper: null}],
+  ['1m', {type: duration, lower: 'now() - 300m', upper: null}],
+  [
+    '1m5s',
+    {
+      type: duration,
+      lower: 'now() - 300m1500s',
+      upper: null,
+    },
+  ],
 ]
 
 test.each(TESTS)('getCheckVisTimeRange(%s)', (input, expected) => {