+++ ui/src/dataLoaders/reducers/telegrafEditor.ts
@@ -4,9 +4,8 @@ import {
   ActivePluginAction,
   EditorAction,
 } from 'src/dataLoaders/actions/telegrafEditor'
-type TelegrafEditorPluginType =
+export type TelegrafEditorPluginType =
   | 'system'
-  | 'bundle'
   | 'input'
   | 'output'
   | 'processor'
@@ -14,24 +13,26 @@ type TelegrafEditorPluginType =
   | 'display'
 type TelegrafEditorPluginName = string
 
-interface TelegrafEditorBasicPlugin {
+export interface TelegrafEditorBasicPlugin {
   name: TelegrafEditorPluginName
   description: string
   code: string
   type: TelegrafEditorPluginType
 }
 
-interface TelegrafEditorBundlePlugin {
+export interface TelegrafEditorBundlePlugin {
   name: TelegrafEditorPluginName
   description: string
-  type: string
+  type: 'bundle'
   include: Array<TelegrafEditorPluginName>
 }
-export type TelegrafEditorPluginState = Array<
-  TelegrafEditorBasicPlugin | TelegrafEditorBundlePlugin
->
 
-interface TelegrafEditorActivePlugin {
+export type TelegrafEditorPlugin =
+  | TelegrafEditorBasicPlugin
+  | TelegrafEditorBundlePlugin
+export type TelegrafEditorPluginState = Array<TelegrafEditorPlugin>
+
+export interface TelegrafEditorActivePlugin {
   name: string
   type: TelegrafEditorPluginType
   line: number
@@ -40,7 +41,8 @@ interface TelegrafEditorActivePlugin {
 export type TelegrafEditorActivePluginState = Array<TelegrafEditorActivePlugin>
 
 type TelegrafEditorMode = 'adding' | 'indexing'
-interface TelegrafEditorState {
+
+export interface TelegrafEditorState {
   mode: TelegrafEditorMode
   bucket: Bucket | null
   text: string
@@ -52,7 +54,8 @@ const INITIAL_PLUGINS: TelegrafEditorPluginState = [
     name: 'cpu',
     type: 'input',
     description: 'watch your cpu yo',
-    code: `[[inputs.cpu]]
+    code: `
+[[inputs.cpu]]
   ## Whether to report per-cpu stats or not
   percpu = true
   ## Whether to report total system cpu stats or not
@@ -67,7 +70,8 @@ const INITIAL_PLUGINS: TelegrafEditorPluginState = [
     name: 'disk',
     type: 'input',
     description: 'watch your disks yo',
-    code: `[[inputs.disk]]
+    code: `
+[[inputs.disk]]
 ## By default stats will be gathered for all mount points.
 ## Set mount_points will restrict the stats to only the specified mount points.
 # mount_points = ["/"]
@@ -79,21 +83,24 @@ ignore_fs = ["tmpfs", "devtmpfs", "devfs", "overlay", "aufs", "squashfs"]
     name: 'diskio',
     type: 'input',
     description: 'watch your diskio yo',
-    code: `[[inputs.diskio]]
+    code: `
+[[inputs.diskio]]
 `,
   },
   {
     name: 'memory',
     type: 'input',
     description: 'watch your memory yo',
-    code: `[[inputs.mem]]
+    code: `
+[[inputs.mem]]
 `,
   },
   {
     name: 'network',
     type: 'input',
     description: 'watch your network yo',
-    code: `[[inputs.net]]
+    code: `
+[[inputs.net]]
 `,
   },
   {
@@ -106,10 +113,11 @@ ignore_fs = ["tmpfs", "devtmpfs", "devfs", "overlay", "aufs", "squashfs"]
     name: 'kubernetes',
     type: 'input',
     description: 'watch your cluster yo',
-    code: `[[inputs.kubernetes]]
-## URL for the kubelet
-## exp: http://1.1.1.1:10255
-url = "http://url"
+    code: `
+[[inputs.kubernetes]]
+  ## URL for the kubelet
+  ## exp: http://1.1.1.1:10255
+  url = "http://url"
 `,
   },
   {
@@ -176,7 +184,8 @@ omit_hostname = false
     name: 'influxdb_v2',
     type: 'output',
     description: 'output to the cloud',
-    code: `[[outputs.influxdb_v2]]
+    code: `
+[[outputs.influxdb_v2]]
 ## The URLs of the InfluxDB cluster nodes.
 ##
 ## Multiple URLs can be specified for a single cluster, only ONE of the
@@ -222,7 +231,7 @@ export function pluginsReducer(
 }
 
 export function activePluginsReducer(
-  state = [],
+  state: TelegrafEditorActivePluginState = [],
   action: ActivePluginAction
 ): TelegrafEditorActivePluginState {
   switch (action.type) {