+++ ui/src/timeMachine/components/Queries.tsx
@@ -7,9 +7,7 @@ import TimeMachineFluxEditor from 'src/timeMachine/components/TimeMachineFluxEdi
 import CSVExportButton from 'src/shared/components/CSVExportButton'
 import TimeMachineQueriesSwitcher from 'src/timeMachine/components/QueriesSwitcher'
 import TimeMachineRefreshDropdown from 'src/timeMachine/components/RefreshDropdown'
-import TimeRangeDropdown, {
-  RangeType,
-} from 'src/shared/components/TimeRangeDropdown'
+import TimeRangeDropdown from 'src/shared/components/TimeRangeDropdown'
 import TimeMachineQueryBuilder from 'src/timeMachine/components/QueryBuilder'
 import SubmitQueryButton from 'src/timeMachine/components/SubmitQueryButton'
 import RawDataToggle from 'src/timeMachine/components/RawDataToggle'
@@ -94,15 +92,12 @@ class TimeMachineQueries extends PureComponent<Props> {
     )
   }
 
-  private handleSetTimeRange = (
-    timeRange: TimeRange,
-    rangeType: RangeType = RangeType.Relative
-  ) => {
+  private handleSetTimeRange = (timeRange: TimeRange) => {
     const {autoRefresh, onSetAutoRefresh, onSetTimeRange} = this.props
 
     onSetTimeRange(timeRange)
 
-    if (rangeType === RangeType.Absolute) {
+    if (timeRange.type === 'custom') {
       onSetAutoRefresh({...autoRefresh, status: AutoRefreshStatus.Disabled})
       return
     }