+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features
 
 1. [15836](https://github.com/influxdata/influxdb/pull/16077): Add stacked line layer option to graphs
+1. [16094](https://github.com/influxdata/influxdb/pull/16094): Annotate log messages with trace ID, if available
+1. [16187](https://github.com/influxdata/influxdb/pull/16187): Bucket create to accept an org name flag
+1. [16158](https://github.com/influxdata/influxdb/pull/16158): Add trace ID response header to query endpoint
 
 ### Bug Fixes
 
@@ -16,8 +19,10 @@
 1. [16101](https://github.com/influxdata/influxdb/pull/16101): Gracefully handle invalid user-supplied JSON
 1. [16105](https://github.com/influxdata/influxdb/pull/16105): Fix crash when loading queries built using Query Builder
 1. [16112](https://github.com/influxdata/influxdb/pull/16112): Create cell view properties on dashboard creation
+1. [16144](https://github.com/influxdata/influxdb/pull/16144): Scrollbars are dapper and proper
 1. [16172](https://github.com/influxdata/influxdb/pull/16172): Fixed table ui threshold colorization issue where setting thresholds would not change table UI
 1. [16194](https://github.com/influxdata/influxdb/pull/16194): Fixed windowPeriod issue that stemmed from webpack rules
+1. [16175](https://github.com/influxdata/influxdb/pull/16175): Added delete functionality to note cells so that they can be deleted
 
 ### UI Improvements